// src/lib/api/generated/model/deleteDocumentResponse.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Myelin Base RAG API
 * API documentation for Myelin Base RAG backend
 * OpenAPI spec version: 1.0.0
 */
import type { DocumentResponse } from "./documentResponse";

export interface DeleteDocumentResponse {
  document: DocumentResponse;
}