// src/lib/api/generated/model/getDocumentResponse.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Myelin Base RAG API
 * API documentation for Myelin Base RAG backend
 * OpenAPI spec version: 1.0.0
 */
import type { DocumentResponse } from "./documentResponse";

export interface GetDocumentResponse {
  document: DocumentResponse;
}